+++ providers/implementations/ciphers/build.info
@@ -109,9 +109,9 @@ SOURCE[$AES_GOAL]=\
         cipher_aes_xts.c cipher_aes_xts_hw.c \
         cipher_aes_gcm.c cipher_aes_gcm_hw.c \
         cipher_aes_gcm_hw_ppc.c \
-        cipher_aes_gcm_hw_s390x.c cipher_aes_gcm_hw_t4.c \
+        cipher_aes_gcm_hw_t4.c \
         cipher_aes_ccm.c cipher_aes_ccm_hw.c \
-        cipher_aes_ccm_hw_s390x.c cipher_aes_ccm_hw_t4.c \
+        cipher_aes_ccm_hw_t4.c \
         cipher_aes_wrp.c \
         cipher_aes_cbc_hmac_sha.c \
         cipher_aes_cbc_hmac_sha256_hw.c cipher_aes_cbc_hmac_sha1_hw.c \
