+++ providers/implementations/ciphers/cipher_aes_xts_hw.c
@@ -142,144 +142,6 @@ static const PROV_CIPHER_HW *ossl_prov_cipher_hw_aes_xts_t4()
     return NULL;
 }
 
-#elif defined(AES_XTS_S390X)
-
-int s390x_aes_xts_cipher_stream(PROV_AES_XTS_CTX *xctx,
-    unsigned char *out, size_t *outl,
-    const unsigned char *in, size_t inl)
-{
-    S390X_KM_XTS_PARAMS *km = &xctx->plat.s390x.param.km;
-    unsigned char *param = (unsigned char *)km + xctx->plat.s390x.offset;
-    unsigned int fc = xctx->plat.s390x.fc;
-    unsigned char tmp[2][AES_BLOCK_SIZE];
-    unsigned char nap_n1[AES_BLOCK_SIZE];
-    unsigned char drop[AES_BLOCK_SIZE];
-    size_t len_incomplete, len_complete;
-
-    len_incomplete = inl % AES_BLOCK_SIZE;
-    len_complete = (len_incomplete == 0) ? inl : (inl / AES_BLOCK_SIZE - 1) * AES_BLOCK_SIZE;
-
-    if (len_complete > 0)
-        s390x_km(in, len_complete, out, fc, param);
-    if (len_incomplete == 0)
-        goto out;
-
-    memcpy(tmp, in + len_complete, AES_BLOCK_SIZE + len_incomplete);
-    /* swap NAP for decrypt */
-    if (fc & S390X_DECRYPT) {
-        memcpy(nap_n1, km->nap, AES_BLOCK_SIZE);
-        s390x_km(tmp[0], AES_BLOCK_SIZE, drop, fc, param);
-    }
-    s390x_km(tmp[0], AES_BLOCK_SIZE, tmp[0], fc, param);
-    if (fc & S390X_DECRYPT)
-        memcpy(km->nap, nap_n1, AES_BLOCK_SIZE);
-
-    memcpy(tmp[1] + len_incomplete, tmp[0] + len_incomplete,
-        AES_BLOCK_SIZE - len_incomplete);
-    s390x_km(tmp[1], AES_BLOCK_SIZE, out + len_complete, fc, param);
-    memcpy(out + len_complete + AES_BLOCK_SIZE, tmp[0], len_incomplete);
-
-    /* do not expose temporary data */
-    OPENSSL_cleanse(tmp, sizeof(tmp));
-out:
-    memcpy(xctx->base.iv, km->tweak, AES_BLOCK_SIZE);
-    *outl = inl;
-
-    return 1;
-}
-
-static int cipher_hw_aes_xts_s390x_initkey(PROV_CIPHER_CTX *ctx,
-    const unsigned char *key, size_t keylen)
-{
-    PROV_AES_XTS_CTX *xctx = (PROV_AES_XTS_CTX *)ctx;
-    S390X_KM_XTS_PARAMS *km = &xctx->plat.s390x.param.km;
-    unsigned int fc, offs;
-    unsigned int dec = 0;
-    int supported = 0;
-
-    switch (keylen) {
-    case 128 / 8 * 2:
-        fc = S390X_XTS_AES_128_MSA10;
-        offs = 32;
-        break;
-    case 256 / 8 * 2:
-        fc = S390X_XTS_AES_256_MSA10;
-        offs = 0;
-        break;
-    default:
-        fc = 0;
-        break;
-    }
-
-    if (fc != 0)
-        supported = (OPENSSL_s390xcap_P.km[1] && S390X_CAPBIT(fc));
-    if (!supported) {
-        xctx->plat.s390x.fc = 0;
-        xctx->plat.s390x.offset = 0;
-        return 0;
-    }
-
-    if (xctx->base.iv_set) {
-        if (xctx->base.ivlen > sizeof(km->tweak)) {
-            ERR_raise(ERR_LIB_PROV, PROV_R_INVALID_IV_LENGTH);
-            return 0;
-        }
-        memcpy(km->tweak, xctx->base.iv, xctx->base.ivlen);
-        xctx->plat.s390x.iv_set = 1;
-    }
-
-    if (key != NULL) {
-        memcpy(km->key + offs, key, keylen);
-        xctx->plat.s390x.key_set = 1;
-    }
-
-    if (xctx->base.enc == 0)
-        dec = S390X_DECRYPT;
-
-    xctx->plat.s390x.fc = fc | dec;
-    xctx->plat.s390x.offset = offs;
-
-    memset(km->nap, 0, sizeof(km->nap));
-    km->nap[0] = 0x1;
-
-    return 1;
-}
-
-static void cipher_hw_aes_xts_s390x_copyctx(PROV_CIPHER_CTX *dst,
-    const PROV_CIPHER_CTX *src)
-{
-    PROV_AES_XTS_CTX *sctx = (PROV_AES_XTS_CTX *)src;
-    PROV_AES_XTS_CTX *dctx = (PROV_AES_XTS_CTX *)dst;
-
-    *dctx = *sctx;
-    dctx->xts.key1 = NULL;
-    dctx->xts.key2 = NULL;
-}
-
-static const PROV_CIPHER_HW aes_xts_s390x = {
-    cipher_hw_aes_xts_s390x_initkey,
-    NULL,
-    cipher_hw_aes_xts_s390x_copyctx
-};
-
-static const PROV_CIPHER_HW *ossl_prov_cipher_hw_aes_xts_s390x(size_t keybits)
-{
-    switch (keybits) {
-    case (128 * 2):
-        if (OPENSSL_s390xcap_P.km[1] && S390X_CAPBIT(S390X_XTS_AES_128_MSA10))
-            return &aes_xts_s390x;
-        break;
-    case (256 * 2):
-        if (OPENSSL_s390xcap_P.km[1] && S390X_CAPBIT(S390X_XTS_AES_256_MSA10))
-            return &aes_xts_s390x;
-        break;
-    default:
-        break;
-    }
-
-    return NULL;
-}
-
 #endif
 
 static const PROV_CIPHER_HW aes_generic_xts = {
